Prevent premature temporary file deletion
To open a file directly from the 7-Zip file manager, either by using a
double click or drag & drop, the 7zFM application creates a temporary
directory to allow the external application to access the file.
The problem is that the file is deleted too soon, and often the external
application has no time to read it.

This can be avoided by deleting the file only when the 7zFM is closed.

More details:
https://sourceforge.net/p/sevenzip/discussion/45797/thread/3b8f3e73/